+++ src/docker/ids.nim
@@ -188,14 +188,17 @@ proc parseImage*(name: string, defaultTag = "latest"): DockerImage =
     let (repo, tag) = image.splitBy(":", defaultTag)
     return (repo, tag, digest)
 
-proc parseImages*(names: seq[string]): seq[DockerImage] =
+proc parseImages*(names: seq[string], defaultTag = "latest"): seq[DockerImage] =
   result = @[]
   for name in names:
     if name != "":
-      result.add(parseImage(name))
+      result.add(parseImage(name, defaultTag = defaultTag))
 
 proc withTag*(self: DockerImage, tag: string): DockerImage =
   return (self.repo, tag, self.digest)
+
+proc isPinned*(self: DockerImage): bool =
+  return self.digest != "" or self.repo == "scratch"
 
 proc withDigest*(self: DockerImage, digest: string): DockerImage =
   return (self.repo, self.tag, digest.extractDockerHash())