+++ src/docker/image.nim
@@ -7,7 +7,7 @@
 
 ## utilities for interacting with docker images
 
-import ".."/[config]
+import ".."/[config, util]
 import "."/[exe, ids, inspect, manifest, dockerfile]
 
 proc pullImage*(name: string) =
@@ -25,6 +25,46 @@ proc pullImage*(name: string) =
       stdout & " " & stderr,
     )
 
+proc fetchManifestForImage*(image: DockerImage, platforms: seq[DockerPlatform]): DockerManifest =
+  trace("docker: fetching manifest list for: " & $image & " " & $($platforms))
+  try:
+    if len(platforms) > 1:
+      raise newException(ValueError, "multi-platform build. cannot inspect local image for manifest digest")
+    # if image is present locally, honor its digest
+    let
+      local   = inspectImageJson(image.asRepoRef(), platforms[0])
+      digests = parseImages(local{"RepoDigests"}.getStrElems(), defaultTag = image.tag)
+    var
+      sameRepo     = newSeq[DockerImage]()
+      sameRegistry = newSeq[DockerImage]()
+      other        = newSeq[DockerImage]()
+    for remote in digests:
+      # take precedence of locally pulled image for the same repo. e.g:
+      # docker pull alpine
+      # echo FROM alpine | docker build -f - .
+      if remote.repo == image.repo:
+        sameRepo.add(remote)
+      # attempt all get manifest from renamed refs but in same registry.
+      # as it is the same registry having same digest is much more likely
+      # e.g:
+      # docker pull alpine
+      # docker tag alpine foo
+      # echo FROM foo | docker build -f - .
+      elif remote.registry == image.registry:
+        sameRegistry.add(remote)
+      # finally try all other registries
+      else:
+        other.add(remote)
+    for remote in sameRepo & sameRegistry & other:
+      try:
+        return fetchListOrImageManifest(remote, platforms)
+      except:
+        trace("docker: " & getCurrentExceptionMsg())
+    raise newException(ValueError, "Could not find manifest list for " & $image)
+  except:
+    trace("docker: " & getCurrentExceptionMsg())
+    return fetchListOrImageManifest(image, platforms)
+
 proc fetchImageOrManifestConfig(image: DockerImage, platform: DockerPlatform): JsonNode =
   trace("docker: fetching config for: " & $image & " " & $platform)
   try:
@@ -33,7 +73,7 @@ proc fetchImageOrManifestConfig(image: DockerImage, platform: DockerPlatform): J
     if hasBuildX():
       return fetchImageManifest(image, platform).config.json{"config"}
     else:
-      trace("docker: buildx is missing. pulling image locally for inspection")
+      trace("docker: buildx is missing. pulling image locally for inspection: " & getCurrentExceptionMsg())
       pullImage(image.asRepoRef())
       return inspectImageJson(image.asRepoRef(), platform){"Config"}