+++ tests/functional/test_docker.py
@@ -3,6 +3,7 @@
 # This file is part of Chalk
 # (see https://crashoverride.com/docs/chalk)
 import platform
+import re
 import shutil
 import time
 from contextlib import ExitStack
@@ -351,31 +352,76 @@ def test_recursion_wrapping(chalk: Chalk, random_hex: str):
     assert run
 
 
-def test_base_images(chalk: Chalk):
+def test_base_images(chalk: Chalk, random_hex: str):
+    # most of the images below are manifest lists
+    # so we create a dummy one which is guaranteed to be regular image manifest
+    image = f"{REGISTRY}/{random_hex}_image"
+    Docker.build(
+        tag=image,
+        content=Docker.dockerfile(
+            """
+            FROM alpine
+            CMD /true
+            """
+        ),
+        push=True,
+        buildx=True,
+    )
+
     _, result = chalk.docker_build(
-        dockerfile=DOCKERFILES / "valid" / "bases" / "Dockerfile",
+        content=Docker.dockerfile(
+            f"""
+            ARG BASE=two
+
+            FROM alpine as one
+
+            FROM ubuntu:24.04 as two
+            COPY --from=docker /usr/local/bin/docker /docker
+            COPY --from=busybox:latest /bin/busybox /busybox
+
+            FROM busybox@sha256:9ae97d36d26566ff84e8893c64a6dc4fe8ca6d1144bf5b87b2b85a32def253c7 as three
+
+            FROM nginx:1.27.0@sha256:97b83c73d3165f2deb95e02459a6e905f092260cd991f4c4eae2f192ddb99cbe as four
+
+            FROM one as five
+            COPY --from=nginx:1.27.0@sha256:97b83c73d3165f2deb95e02459a6e905f092260cd991f4c4eae2f192ddb99cbe /usr/sbin/nginx /nginx
+            COPY --from=one /bin/sh /sh
+
+            FROM scratch as six
+
+            FROM {image}:latest as seven
+
+            FROM $BASE
+            COPY --from=four /usr/sbin/nginx /nginx
+            """
+        ),
     )
     assert result.mark.has(
         DOCKER_TARGET="",
-        DOCKER_BASE_IMAGE="ubuntu:24.04",
+        DOCKER_BASE_IMAGE=re.compile(r"ubuntu:24.04@sha256:"),
         DOCKER_BASE_IMAGE_REPO="ubuntu",
         DOCKER_BASE_IMAGE_TAG="24.04",
+        DOCKER_BASE_IMAGE_DIGEST=ANY,
         DOCKER_BASE_IMAGES={
             "one": {
-                "from": "alpine",
-                "name": "alpine",
+                "from": re.compile("alpine@sha256:"),
+                "name": re.compile("alpine@sha256:"),
                 "repo": "alpine",
+                "tag": MISSING,
+                "digest": ANY,
             },
             "two": {
-                "from": "ubuntu:24.04",
-                "name": "ubuntu:24.04",
+                "from": re.compile("ubuntu:24.04@sha256:"),
+                "name": re.compile("ubuntu:24.04@sha256:"),
                 "repo": "ubuntu",
                 "tag": "24.04",
+                "digest": ANY,
             },
             "three": {
                 "from": "busybox@sha256:9ae97d36d26566ff84e8893c64a6dc4fe8ca6d1144bf5b87b2b85a32def253c7",
                 "name": "busybox@sha256:9ae97d36d26566ff84e8893c64a6dc4fe8ca6d1144bf5b87b2b85a32def253c7",
                 "repo": "busybox",
+                "tag": MISSING,
                 "digest": "9ae97d36d26566ff84e8893c64a6dc4fe8ca6d1144bf5b87b2b85a32def253c7",
             },
             "four": {
@@ -387,14 +433,31 @@ def test_base_images(chalk: Chalk):
             },
             "five": {
                 "from": "one",
-                "name": "alpine",
+                "name": re.compile("alpine@sha256:"),
                 "repo": "alpine",
+                "tag": MISSING,
+                "digest": ANY,
+            },
+            "six": {
+                "from": "scratch",
+                "name": "scratch",
+                "repo": "scratch",
+                "tag": MISSING,
+                "digest": MISSING,
+            },
+            "seven": {
+                "from": re.compile(f"{image}:latest@sha256:"),
+                "name": re.compile(f"{image}:latest@sha256:"),
+                "repo": image,
+                "tag": "latest",
+                "digest": ANY,
             },
             "": {
                 "from": "two",
-                "name": "ubuntu:24.04",
+                "name": re.compile("ubuntu:24.04@sha256:"),
                 "repo": "ubuntu",
                 "tag": "24.04",
+                "digest": ANY,
             },
         },
         DOCKER_COPY_IMAGES={
@@ -403,6 +466,8 @@ def test_base_images(chalk: Chalk):
                     "from": "docker",
                     "name": "docker",
                     "repo": "docker",
+                    "tag": MISSING,
+                    "digest": MISSING,
                     "src": ["/usr/local/bin/docker"],
                     "dest": "/docker",
                 },
@@ -411,6 +476,7 @@ def test_base_images(chalk: Chalk):
                     "name": "busybox:latest",
                     "repo": "busybox",
                     "tag": "latest",
+                    "digest": MISSING,
                     "src": ["/bin/busybox"],
                     "dest": "/busybox",
                 },
@@ -427,8 +493,10 @@ def test_base_images(chalk: Chalk):
                 },
                 {
                     "from": "one",
-                    "name": "alpine",
+                    "name": re.compile("alpine@sha256:"),
                     "repo": "alpine",
+                    "tag": MISSING,
+                    "digest": ANY,
                     "src": ["/bin/sh"],
                     "dest": "/sh",
                 },